- hardware version as a parameter to select template - template no longer required argument, but can be set with --template - add a config file (we need to store the PREFIX, but can be used for other settings) - add --ovf option - update README.md - use --format-ustar as required in the OVF format spec

+++ ova/mkova.sh
@@ -15,8 +15,27 @@
 # specific language governing permissions and limitations under the License.
 # ================================================================================
 
+[ -f /etc/open-vmdk.conf ] && . /etc/open-vmdk.conf
+
+[ ! -n "$PREFIX" ] && PREFIX=/usr
+[ ! -n "$DATADIR" ] && DATADIR=${PREFIX}/share/open-vmdk
+# default template should be the latest:
+[ ! -n "$OVFTEMPL" ] && OVFTEMPL=${DATADIR}/template-hw20.ovf
+
+[ ! -n "$NUM_CPUS" ] && NUM_CPUS=2
+[ ! -n "$MEM_SIZE" ] && MEM_SIZE=1024
+[ ! -n "$FIRMWARE" ] && FIRMWARE="efi"
+
 usage() {
-    echo "Usage: $0 [-n|--num-cpus <num_cpus>] [-m|--mem-size <mem_size>] [-f|--firmware efi|bios] ova_name path_to_ovf_template disk1.vmdk [disk2.vmdk [...]]"
+    echo "Usage: $0"
+    echo "          [-n|--num-cpus <num_cpus>]"
+    echo "          [-m|--mem-size <mem_size>]"
+    echo "          [-f|--firmware efi|bios]"
+    echo "          [--hw <version>]"
+    echo "          [--template <path_to_template>]"
+    echo "          [--ovf]"
+    echo "          ova_name disk1.vmdk [disk2.vmdk [...]]"
+    echo ""
     echo "Create an OVA template from VMDK files"
     echo ""
     echo "Environment variables can be used to change OVA settings:"
@@ -25,11 +44,7 @@ usage() {
     echo "* FIRMWARE: The firmware of the OVA template: efi or bios. Default value is efi."
 }
 
-[ ! -n "$NUM_CPUS" ] && NUM_CPUS=2
-[ ! -n "$MEM_SIZE" ] && MEM_SIZE=1024
-[ ! -n "$FIRMWARE" ] && FIRMWARE="efi"
-
-OPTS=$(getopt -o n:m:f: --long num-cpus:,mem-size:,firmware: -n $0 -- "$@")
+OPTS=$(getopt -o n:m:f: --long num-cpus:,mem-size:,firmware:,hw:,template:,ovf -n $0 -- "$@")
 if [ $? != 0 ] ; then
     usage
     echo "Terminating." >&2
@@ -38,27 +53,37 @@ fi
 
 eval set -- "$OPTS"
 
+ovftempl=${OVFTEMPL}
+ovf=false
+
 while true; do
     case "$1" in
         -n | --num-cpus) NUM_CPUS=${2} ; shift 2 ;;
         -m | --mem-size) MEM_SIZE=${2} ; shift 2 ;;
         -f | --firmware) FIRMWARE=${2} ; shift 2 ;;
+        --hw) ovftempl=${DATADIR}/template-hw${2}.ovf ; shift 2 ;;
+        --template) ovftempl=${2} ; shift 2 ;;
+        --ovf) ovf="true" ; shift ;;
         --) shift; break ;;
         *) break ;;
     esac
 done
 
-if [ "$#" -lt 3 ] ; then
-    echo "need at least 3 arguments" >&2
+if [ "$#" -lt 2 ] ; then
+    echo "need at least 2 arguments" >&2
     usage
     echo "Terminating." >&2
     exit 1
 fi
 
 name=$1
-ovftempl=$2
 shift
-shift
+# for backwards compatibility:
+if [ "${1: -4}" == ".ovf" ] ; then
+    echo "Specifying the template as an argument is deprecated. Please use the --template or the --hw option."
+    ovftempl=$1
+    shift
+fi
 vmdks=$@
 vmdks_num=$#
 
@@ -167,10 +192,21 @@ done
 # Get the sha checksum of the ovf file
 echo "SHA${sha_alg}(${name}.ovf)= $(sha${sha_alg}sum $TMPDIR/${name}.ovf | cut -d' ' -f1)" >> $TMPDIR/${name}.mf
 
-pushd $TMPDIR
-tar cf ../${name}.ova *.ovf *.mf *.vmdk
-popd
+if [ ${ovf} == "true" ] ; then
+    if [ -d ${name} ] ; then
+        echo "directory ${name} exists" >&2
+        echo "Terminating." >&2
+        exit 3
+    fi
+    mv $TMPDIR ${name}
+    echo "Completed to create ovf files in directory ${name}"
+else
+    pushd $TMPDIR
+    tar --format=ustar -cf ../${name}.ova *.ovf *.mf *.vmdk
+    popd
 
-echo "Completed to create ${name}.ova"
+    echo "Completed to create ${name}.ova"
 
-rm -rf $TMPDIR
+    rm -rf $TMPDIR
+fi
+
